Very Basic Preview Function
Implemented a very basic preview function for the shapes.

Also fixed bug of edges accidently deleting when clicking a node, was a very simple fix to give nodes hitboxes priority.

Found new bug when testing the square. My currently implementation of validating the solution isn't strong enough to handle cases where the player links the square as a cross.