Fixed nodeID bug.
Instead of using a static variable (technically it was a global variable) on the node and increasing the value by 1 each time a new node is created, each graph now maintains its own ID counter. This was an oversight from me, and would've created issues if someone else uses the class.