Character hair update
Rigged the character model that was broken up into parts for customisation so that changes to hair etc can be reflected in game.

Model is not perfect and will need work. Just a proof of concept at this stage but is now animated again.

Few things still to fix:
Assigning colour to items and saving that selection