New Feature Shape Preview Panel
Implemented a shape preview panel which aids the user in searching for a shape within a level. This panel shows the current shape on the level and stops user input when open.

I also refactored levels to use JSON to load the initial data before procedurally generating the stars. This just defines the shape/s for a level and the rotation.